requirements.txt
# PYTHON==3.7.10 #
python-dotenv
aiogram
openai==0.28